+++ biblioteca-4/administrador/calendario.php
@@ -133,7 +133,7 @@ $consulta = "SELECT COUNT(*)
 $result = $pdo->prepare($consulta);
 $result->execute([":prestado" => $fecha]);
 if (!$result) {
-    print "    <p class=\"aviso\">Error en la consulta.</p>\n";
+    print "    <p class=\"aviso-error\">Error en la consulta.</p>\n";
 } elseif ($result->fetchColumn() == 0) {
     print "    <p>Haga clic en los días del mes con enlaces para ver los préstamos realizados en ese día.</p>\n";
 } else {
@@ -152,7 +152,7 @@ if (!$result) {
     $result = $pdo->prepare($consulta);
     $result->execute([":prestado" => $fecha]);
     if (!$result) {
-        print "    <p class=\"aviso\">Error en la consulta.</p>\n";
+        print "    <p class=\"aviso-error\">Error en la consulta.</p>\n";
     } else {
         print "    <p>Listado completo de registros:</p>\n";
         print "\n";