+++ biblioteca-4/tablas/db-personas/modificar-2.php
@@ -35,7 +35,7 @@ $consulta = "SELECT *
 $result = $pdo->prepare($consulta);
 $result->execute([":id" => $id]);
 if (!$result) {
-    print "    <p class=\"aviso\">Error en la consulta.</p>\n";
+    print "    <p class=\"aviso-error\">Error en la consulta.</p>\n";
 } else {
     $valor = $valor = $result->fetch(PDO::FETCH_ASSOC);
     print "    <form action=\"modificar-3.php\" method=\"$cfg[formMethod]\">\n";