+++ biblioteca-4/tablas/db-personas/modificar-3.php
@@ -41,9 +41,9 @@ $consulta = "UPDATE $db[personas]
              WHERE id=:id";
 $result = $pdo->prepare($consulta);
 if ($result->execute([":nombre" => $nombre, ":apellidos" => $apellidos, ":dni" => $dni, ":id" => $id])) {
-    print "    <p>Registro modificado correctamente.</p>\n";
+    print "    <p class=\"aviso-info\">Registro modificado correctamente.</p>\n";
 } else {
-    print "    <p class=\"aviso\">Error al modificar el registro.</p>\n";
+    print "    <p class=\"aviso-error\">Error al modificar el registro.</p>\n";
 }
 
 $pdo = null;