+++ biblioteca-4/tablas/db-prestamos/modificar-3.php
@@ -45,9 +45,9 @@ $consulta = "UPDATE $db[prestamos]
 $result = $pdo->prepare($consulta);
 if ($result->execute([":id_persona" => $id_persona, ":id_obra" => $id_obra,
     ":prestado" => $prestado, ":devuelto" => $devuelto, ":id" => $id, ])) {
-    print "    <p>Registro modificado correctamente.</p>\n";
+    print "    <p class=\"aviso-info\">Registro modificado correctamente.</p>\n";
 } else {
-    print "    <p class=\"aviso\">Error al modificar el registro.</p>\n";
+    print "    <p class=\"aviso-error\">Error al modificar el registro.</p>\n";
 }
 
 $pdo = null;